requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-huggingface
langchain-chroma
pypdf
pytest
ruff
python-dotenv
streamlit
sentence-transformers